% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP-framework-sub-functions.R
\name{compute_bearing}
\alias{compute_bearing}
\title{Bearing computation}
\usage{
compute_bearing(lat_1, lon_1, lat_2, lon_2)
}
\arguments{
\item{lat_1}{The decimal degree latitude value of coordinate one.}

\item{lon_1}{The decimal degree longitude value of coordinate one.}

\item{lat_2}{The decimal degree latitude value of coordinate two.}

\item{lon_2}{The decimal degree longitude value of coordinate two.}
}
\value{
The absolute bearing or heading angle from coordinate one to coordinate two in degrees.
}
\description{
Computes the bearing or heading angle between a pair of decimal degree geospatial coordinates.
}
\examples{
my_bearing <- compute_bearing(lat_1 = 8.46696, lon_1 = -17.03663, lat_2 = 32.918481, lon_2 = 11.96337)
}
