% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP-framework-sub-functions.R
\name{compute_turning_angle}
\alias{compute_turning_angle}
\title{Turning angle computation}
\usage{
compute_turning_angle(bearing_vec_1, bearing_vec_2)
}
\arguments{
\item{bearing_vec_1}{A vector or numeric value of bearing angles.}

\item{bearing_vec_2}{A vector or numeric value of bearing angles, one observation ahead of bear_vec_1.}
}
\value{
The absolute turning angle or the change in bearing angles of an object in degrees.
}
\description{
Computes the turning angle of an object.
}
\examples{
my_turning_angle <- compute_turning_angle(bearing_vec_1 = 90, bearing_vec_2 = 60)
}
