% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP_framework_sub_functions.R
\name{create_movement_unit_dictionary}
\alias{create_movement_unit_dictionary}
\title{Creates movement unit dictionary}
\usage{
create_movement_unit_dictionary(
  default,
  velocity_label,
  acceleration_label,
  turning_angle_label
)
}
\arguments{
\item{default}{If default is = TRUE the function builds default movement unit dictionary.}

\item{velocity_label}{A character string vector containing velocity descriptors. 
The default argument is c("Walk", "Jog", "Run", "Sprint").}

\item{acceleration_label}{A character string vector containing acceleration descriptors.
The default argument is c("Deceleration", "Neutral", "Acceleration").}

\item{turning_angle_label}{A character string vector containing turning angle descriptors.
The default argument is c("Straight", "Acute-Change", "Large-Change", "Backwards").}
}
\value{
A standardised movement unit dictionary dataframe for use in the SMP framework.
}
\description{
Creates a standardised movement unit dictionary for use in the SMP framework.
}
\examples{
my_movement_dictionary <- create_movement_unit_dictionary(default = TRUE)
}
