% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP-framework-sub-functions.R
\name{hierarchical_cluster_reassignment}
\alias{hierarchical_cluster_reassignment}
\title{Hierarchical cluster reassignment function}
\usage{
hierarchical_cluster_reassignment(
  movement_unit_subsequence_matrix,
  hierarchical_cluster_assignment_df
)
}
\arguments{
\item{movement_unit_subsequence_matrix}{A Levenshtein distance matrix of a participants movement unit subsequences.}

\item{hierarchical_cluster_assignment_df}{A dataframe composed of a participants movement unit subsequences & hierarchical cluster groups.}
}
\value{
A dataframe containing the supplied movement unit subsequences & their newly associated cluster groups.
}
\description{
Reprocesses hierarchical cluster results by identifying single element clusters and
reassigning the single elements to the next most similar cluster.
}
\examples{
my_reassigned_clusters <- hierarchical_cluster_reassignment(movement_unit_subsequence_matrix = player_matrix,
                                                            hierarchical_cluster_assignment_df = player_hclust_df)
}
