% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP_framework_sub_functions.R
\name{identify_frequent_SMP}
\alias{identify_frequent_SMP}
\title{Identify frequent SMP}
\usage{
identify_frequent_SMP(movement_unit_subsequences, number_of_clusters)
}
\arguments{
\item{movement_unit_subsequences}{A dataframe containing the movement unit subsequences and corresponding hcluster group number.}

\item{number_of_clusters}{The number of unique cluster groups contained in the movement unit subsequences dataframe.}
}
\value{
A dataframe containing frequent SMP identified by the LCS algorithm and summary statistics.
}
\description{
Applies the longest common subsequence algorithm to the movement unit subsequences to identify
the frequent SMP within each cluster group.
}
\examples{
number_of_clusters <- length(unique(hierarchical_cluster_reassignment_df$Group))
frequent_smp_df <- identify_frequent_SMP(movement_unit_subsequences = hierarchical_cluster_reassignment_df, 
                                         number_of_clusters = number_of_clusters)
}
