% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP_framework_sub_functions.R
\name{identify_index_ranges}
\alias{identify_index_ranges}
\title{Identify index ranges}
\usage{
identify_index_ranges(velocity_vector, threshold_value)
}
\arguments{
\item{velocity_vector}{A participants velocity vector time-series.}

\item{threshold_value}{A numeric value representing the change point velocity.}
}
\value{
The start and end index values of the movement unit subsequences.
}
\description{
Identifies the movement unit subsequence index's by delineating
the velocity vector time-series using the user selected threshold value.
}
\examples{
my_index_data <- identify_index_ranges(velocity_vector = player_df$Velocity, 
                                       threshold_value = 1.20)
}
