% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP_framework_sub_functions.R
\name{read_gps_files}
\alias{read_gps_files}
\title{Read GPS files function}
\usage{
read_gps_files(gps_file_names, pattern_to_remove)
}
\arguments{
\item{gps_file_names}{A character vector of the names of GPS files in the named directory.}

\item{pattern_to_remove}{A character string that the user wants removed from the list of file names.}
}
\value{
A list of Catapult 10Hz GPS dataframes.
}
\description{
Reads in Catapult 10Hz GPS files from a source directory and removes file type extensions to create a list of player/ position names.
}
\examples{
my_gps_data_list <- read_gps_files(gps_file_names = list.files(pattern = "*.csv"), pattern_to_remove = "")
}
