% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP_framework_sub_functions.R
\name{remove_string_elements}
\alias{remove_string_elements}
\title{Removes string elements}
\usage{
remove_string_elements(file_names, pattern)
}
\arguments{
\item{file_names}{A character string which represents a file name.}

\item{pattern_to_remove}{A character string that the user wants removed.}
}
\value{
A character string with chosen elements removed.
}
\description{
Removes chosen string element or pattern from a list of file names.
}
\examples{
my_string <- remove_string_elements(file_names = "participant_one.csv", pattern_to_remove = ".csv")
}
