% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP_framework_player_level.R
\name{SMP_framework_player_level}
\alias{SMP_framework_player_level}
\title{Player level SMP framework application}
\usage{
SMP_framework_player_level(
  player_df,
  dictionary_df,
  velocity_thresholds,
  acceleration_thresholds,
  threshold_value,
  number_of_clusters,
  movement_unit_subsequence_duration,
  movement_unit_subsequence_length
)
}
\arguments{
\item{player_df}{A Catapult 10Hz GPS dataframe of a singular player for a singular period of any length.}

\item{dictionary_df}{A standardised movement unit dictionary; built using the create_movement_unit_dictionary() function.}

\item{velocity_thresholds}{A numeric vector representing the velocity time series threshold values which
correspond to the velocity descriptors supplied to the movement unit dictionary.}

\item{acceleration_thresholds}{A numeric vector representing the acceleration time series threshold values which
correspond to the acceleration descriptors supplied to the movement unit dictionary. Use c(min(player_df$Acceleration))
for the first observation to avoid errors - see example.}

\item{threshold_value}{A numeric value representing the change point velocity.}

\item{number_of_clusters}{A numeric value with the initial number of clusters the user wants for the Hierarchical cluster
analysis.}

\item{movement_unit_subsequence_duration}{A numeric value representing the minimum number of movement units required for
a movement subsequence to be included in the analysis.}

\item{movement_unit_subsequence_length}{A character string, either "long" or "short" which pertains to the repeated 
movement units within a movement unit subsequence.}
}
\value{
A list of outputs including the original data, the individuals movement unit dictionary with summary stats, 
the frequent SMP and movement unit subsequences with summary stats.
}
\description{
Applies the SMP framework on an individual basis.
}
\examples{
my_SMP_data <- SMP_framework_player_level(player_df = my_df, dictionary_df = my_dictionary,
                                          velocity_thresholds = c(0.00, 1.69, 3.90, 4.99), acceleration_thresholds = c(min(player_df$Acceleration), -0.20, 0.20),
                                          threshold_value = 1.20, number_of_clusters = 25, movement_unit_subsequence_duration = 50,
                                          movement_unit_subsequence_length = "long")
}
