% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMP-framework-team-level.R
\name{SMP_framework_team_level}
\alias{SMP_framework_team_level}
\title{Team level application of the SMP framework}
\usage{
SMP_framework_team_level(
  my_list,
  velocity_thresholds,
  threshold_value,
  movement_unit_subsequence_duration,
  movement_unit_subsequence_length,
  number_of_clusters
)
}
\arguments{
\item{my_list}{A list of Catapult 10Hz GPS dataframes of multiple players for a single match or training drill of any length.}

\item{velocity_thresholds}{A numeric vector representing the velocity time series threshold values which
correspond to the velocity descriptors supplied to the movement unit dictionary.}

\item{threshold_value}{A numeric value representing the change point velocity.}

\item{movement_unit_subsequence_duration}{A numeric value representing the minimum number of movement units required for
a movement subsequence to be included in the analysis.}

\item{movement_unit_subsequence_length}{A character string, either "long" or "short" which pertains to the repeated 
movement units within a movement unit subsequence.}

\item{number_of_clusters}{A numeric value with the initial number of clusters the user wants for the Hierarchical cluster
analysis.}
}
\value{
A list of outputs including the original data, the individuals movement unit dictionary with summary stats, 
the frequent SMP and movement unit subsequences with summary stats for all players.
}
\description{
Applies the SMP framework to a list of Catapult 10Hz GPS dataframes.
}
\examples{
SMP_data <- apply_SMP_framework(my_list = gps_data_list, velocity_thresholds = c(0.00, 1.69, 3.90, 4.99),
                                threshold_value = 1.20, locomotive_event_subsequence_duration = 50, 
                                locomotive_event_subsequence_length = "long", number_of_clusters = 25)
}
